{"message": "feat(cli): SOTA UX overhaul — @file search, !shell, /compact, streaming attribution
New features:
- @ file search: fuzzy file completion triggered by @ in prompt
  (respects .gitignore, 20 results max, 4 depth limit)
- ! shell escape: run shell commands inline (!git status, !npm test)
  with exit code display (✓/✗), 30s timeout, never sent to LLM
- /compact: compress conversation history to free context tokens,
  shows before→after token breakdown
- Content attribution: agent responses render with left accent border (▏)
  for visual distinction from tool output
- Per-response cost: model · elapsed · $cost · tokens · tok/s on one line
- Context gauge: XX% ctx (XX.XK/XXXK) with color-coded thresholds

Bug fixes:
- chain_mode: path traversal via chain name (../../etc/foo)
- plan_mode/chat_session_signals: $EDITOR parsing for multi-word editors
  (code --wait, vim -O)
- git_tools: -- separator on blame/add to prevent option injection
- git_tools: branch name sanitization rejects .. traversal
- watch_mode: silent except→logger.debug for poll/callback failures
- mood_display: deduplicated identical PAD→mood decision trees
- pipe_mode: removed dead markdown/text branch duplication

Cleanup:
- Import order: logger after all imports across 10 files (E402→0)
- Mutable defaults: HookEvent.ALL list→tuple, _CODE_EXTS set→frozenset,
  WATCH_EXTENSIONS/SKIP_DIRS set→frozenset
- Unused imports: removed 5 dead imports from chat_session.py
- Unused return: dropped project_context from _init_session_core
- Unused loop var: j→_j in session_picker.py
- Unicode noqa: RUF001/RUF002 on intentional chars
- _InputState dataclass consolidation (7 module globals → reset-able)
- ChatSession.__init__ split into 4 phase methods (240→28 lines)
- Activity log f-string SQL → parameterized queries
- fix_commands.py shell=True → shlex.split()

Verification: 539/539 CLI tests pass, 0 ruff issues, all py_compile OK"}

+++ aura/cli/commands/heatmap_commands.py
@@ -22,7 +22,7 @@ _SPARKS = "▁▂▃▄▅▆▇█"
 def _spark(value: int, maximum: int, width: int = 16) -> str:
     if maximum <= 0:
         return "." * width
-    filled = max(1, int(round((value / maximum) * width)))
+    filled = max(1, round((value / maximum) * width))
     block = _SPARKS[-1]
     return block * filled + " " * (width - filled)
